% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\docType{data}
\name{datasim}
\alias{datasim}
\title{Random Data for demonstration.}
\format{A List with components X and Y, both of which are matrices.}
\description{
Y   = XB + E, n = 300, p =200, K = 5.
}
\keyword{datasets}
