% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emInd.R
\name{emInd}
\alias{emInd}
\title{vimco}
\usage{
emInd(X, Y, mu0 = NULL, Alpha0 = NULL, sigb0 = NULL, sige0 = NULL,
  maxit = 10^3L, epsStopLogLik = 10^-5)
}
\arguments{
\item{X}{SNP genotypes recoded in terms of additive and dominant components, rows and columns are for n samples and p SNPs, respectively.}

\item{Y}{A phenotype matrix (n by K), each column is a  (quantitavie) phenotype.}

\item{mu0}{Initial values for the variational mean of effect sizes (p by K).}

\item{Alpha0}{Initial values for the variational mean of nonzero indicators (p by K).}

\item{sigb0}{An initial vector for the standard deviations of effect size.}

\item{sige0}{An initial vector for the standard deviations of random errors in all trait.}

\item{maxit}{The maximal iteration in the VBEM algorithm.}

\item{epsStopLoglik}{A tolerance for convergence.}
}
\value{
List of model parameters and local FDR estimates.
}
\description{
Fit Bayesion variable selection regression with variational Bayesion EM algorithm, for each trait seperately.
}
\details{
\code{emInd} fits BVSR for multiple traits, but each trait is modeled seperately.
}
\examples{
 data(datasim)
 fit_Ind = emInd(datasim$X, datasim$Y)
}
