% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emMultiple.R
\name{emMultiple}
\alias{emMultiple}
\title{vimco}
\usage{
emMultiple(X, Y, mu0 = NULL, sigb0 = NULL, Theta0 = NULL,
  Lambda0 = NULL, Alpha0 = NULL, fixlambda = TRUE, maxit = 10^3L,
  epsStopLogLik = 10^-5)
}
\arguments{
\item{X}{SNP genotypes recoded in terms of additive and dominant components, rows and columns are for n samples and p SNPs, respectively.}

\item{Y}{A phenotype matrix (n by K), each column is a  (quantitavie) phenotype.}

\item{mu0}{Initial values for the variational mean of effect sizes (p by K).}

\item{sigb0}{An initial vector (K by 1) for the standard deviation of effect size.}

\item{Theta0}{An inital precision matrix (K by K) for errrors in multiple traits.}

\item{Lambda0}{Inital values for variational means (p by 1) of nonzero indicators.}

\item{Alpha0}{Initial values for the variational mean of nonzero indicators (p by K).}

\item{fixlambda}{a test parameter, always set to be true.}

\item{Maximum}{The maximal iteration in the VBEM algorithm.}

\item{epsStopLoglik}{The tolerance for convergence.}
}
\value{
List of model parameters and local FDR estimates.
}
\description{
Fit variational inference for mulitple correlated traits with variational Bayesion EM algorithm, all traits are jointly analyzed.
}
\details{
\code{emMultiple} fits VIMCO for multiple correlated traits, all traits are modeled jointly.
}
\examples{
 data(datasim)
 fit_Mul = emMultiple(datasim$X, datasim$Y)
}
